// src/app/models/fixture.ts
// Generated by https://quicktype.io

export interface FixtureResponse {
  get:        string;
  parameters: Parameters;
  results:    number;
  paging:     Paging;
  response:   FixtureResult[];
}

export interface Paging {
  current: number;
  total:   number;
}

export interface Parameters {
  league: string;
  team:   string;
  status: StatusEnum;
  last:   string;
}

export enum StatusEnum {
  Ft = "FT",
}

export interface FixtureResult {
  fixture: Fixture;
  league:  League;
  teams:   Goals;
  goals:   GoalGoals;
  score:   Score;
}

export interface Fixture {
  id:        number;
  referee:   string;
  timezone:  Timezone;
  date:      string;
  timestamp: number;
  periods:   Periods;
  venue:     Venue;
  status:    StatusClass;
}

export interface Periods {
  first:  number;
  second: number;
}

export interface StatusClass {
  long:    Long;
  short:   StatusEnum;
  elapsed: number;
}

export enum Long {
  MatchFinished = "Match Finished",
}

export enum Timezone {
  UTC = "UTC",
}

export interface Venue {
  id:   number;
  name: string;
  city: City;
}

export enum City {
  BournemouthDorset = "Bournemouth, Dorset",
  Liverpool = "Liverpool",
  London = "London",
}

export interface Goals {
  home: AwayClass ;
  away: AwayClass ;
}

export interface GoalGoals {
  home: number ;
  away: number ;
}

export interface AwayClass {
  id:     number;
  name:   string;
  logo:   string;
  winner: boolean | null;
}

export interface League {
  id:      number;
  name:    Name;
  country: Country;
  logo:    string;
  flag:    string;
  season:  number;
  round:   string;
}

export enum Country {
  England = "England",
}

export enum Name {
  PremierLeague = "Premier League",
}

export interface Score {
  halftime:  Goals;
  fulltime:  Goals;
  extratime: Goals;
  penalty:   Goals;
}